% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{extractQrteGz}
\alias{extractQrteGz}
\title{Extract data from the parsed .csv.gz file}
\usage{
extractQrteGz(survey.csv.gz, ...)
}
\arguments{
\item{survey.csv.gz}{String. File path to the .csv.gz file.}

\item{...}{Optional arguments to pass to read.table()}
}
\value{
Data frame.
}
\description{
Extract the data from a QRTEngine survey that has already been parsed into a
.csv.gz file. The parsing can be done via the python parser, for which a
function in this package exists - parseQrte(). Or, the online QRTEninge
parser can be used. See: https://parser.qrtengine.com/
}
\examples{
\dontrun{
extractQrteGz("QualtricsParsed/Esoteric_Analogies.csv.gz")
}
}

