% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parseQrte}
\alias{parseQrte}
\title{Parse QRTEngine data.}
\usage{
parseQrte(survey.zip, parser.py, ..., out.dir = "qrteout",
  unlink.out = TRUE)
}
\arguments{
\item{survey.zip}{String. File path to zipped survey csv downloaded from Qualtrics.}

\item{parser.py}{String. File path to the QRTEngine Pythong parser.}

\item{...}{Optional arguments to pass to read.table()}

\item{out.dir}{String. Name (possibly with path to) of directory in which to
store output. Will be created if it doesn't exist. Default = "qrteout"}

\item{unlink.out}{logical. Should output directory and all files be deleted
before function return? Default = TRUE}
}
\value{
Data frame.
}
\description{
Parse QRTEngine data downloaded from Qualtrics as a zip file containing a csv
file. This function requires that Python 2.7 is running on your machine by
default.
}

