% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrtenginer.R
\docType{package}
\name{qrtenginer}
\alias{qrtenginer}
\alias{qrtenginer-package}
\title{qrtenginer: A package for parsing and scoring QRTEngine tests.}
\description{
The qrtenginer package provides two categories of important functions:
parsing and scoring
}
\section{parsing functions}{

The parsing functions ...
}

\section{scoring functions}{

The scoring functions ...
}

\section{compatible surveys}{

A suite of tests have been built to be compatible with the scoring functions
provided by the qrtengine package. There are two purposes for these tests
being available. One is so that researchers can quickly deploy and analyse
data without the burdens of programming. The second is to encourage
researchers to work with a standard test suite so that suitable comparisons
can be made across studies.
}

