% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoreQrte}
\alias{scoreQrte}
\title{Score data from a compatible QRTEngine test}
\usage{
scoreQrte(data, ...)
}
\arguments{
\item{data}{Data Frame of QRTEngine data after it has been parsed - e.g.,
using extractQrteGz() or parseQrte()}

\item{...}{Optional arguements to pass on to various specific scoring
functions.}
}
\value{
Data frame.
}
\description{
This function will score variables for a QRTEngine test from the parsed raw
data. It uses the BlockId column in the parsed data to identify the test and
apply the appropriate variable calculations. It also uses the BlockID to
exclude practice trial data. \cr\cr
To determine the available optional arugments and variables returned by the function: \cr\cr
For Simple Reaction Time test, see ?scoreRts
}

