% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoreRts}
\alias{scoreRts}
\title{Score a set of response times.}
\usage{
scoreRts(rts, min = 100, max = 500)
}
\arguments{
\item{rts}{Numeric vector of response times captured in milliseconds.}

\item{min}{Number. Minumum response time considered to be valid.}

\item{max}{Number. Maximum response time considered to be valid.}
}
\value{
Named vector with proportion of anticipatory resposnes (p.anti),
  proportion of lapses (p.lapse), and mean response time for all responses
  not classified as anticipatory or lapse (mean.rt).
}
\description{
Calculate (1) count of anticipatory responses, (2) count of lapse responses,
(3) count of valid responses, and (4) mean response time for valid responses
(default between 100 and 500 ms) from a vector of response times.
}

