% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{validateRts}
\alias{validateRts}
\title{Convenience function to determine if Response time is valid.}
\usage{
validateRts(rts, min = NULL, max = NULL)
}
\arguments{
\item{rts}{Numeric vector of response times captured in milliseconds.}

\item{min}{Number. Minumum response time considered to be valid.}

\item{max}{Number. Maximum response time considered to be valid.}
}
\description{
Will return a logical vector set to TRUE for any response time falling within
a valid range of times. Can also be to test greater than or less than ONLY (e.g., for
count of anticipatory or lapse responses).
}

